Refactor: Introduce WC_REST_Cacheable trait
Move caching logic to a trait for reuse across controller base classes.
